requirements.txt
django==1.11
python-instagram==1.3.2
djangorestframework==3.8.2
markdown==2.6.11
django-filter==1.1.0
requests==2.18.4
psycopg2-binary==2.7.4
psycopg2==2.7.4
ipython==5.1.0
Pillow==5.1.0
python-printer-escpos==0.0.3